changed location values from float64 to float32

+++ location.go
@@ -3,7 +3,7 @@ package TeleGogo
 // Location represents a point on the map.
 type Location struct {
 	// Longitude as defined by sender
-	Longitude float64 `json:"longitude"`
+	Longitude float32 `json:"longitude"`
 	// Latitude as defined by sender
-	Latitude float64 `json:"latitude"`
+	Latitude float32 `json:"latitude"`
 }
